refactor: spark_utils -> utils
