+++ ds_toolbox/ml/evaluator.py
@@ -5,7 +5,7 @@ import pyspark
 from pyspark.ml.evaluation import BinaryClassificationEvaluator
 import pandas as pd
 
-from ds_toolbox.spark_utils import start_local_spark
+from ds_toolbox.utils import start_local_spark
 
 @typechecked
 def binary_classifier_metrics(
