+++ tests/conftest.py
@@ -2,7 +2,7 @@ import pickle
 from numpy import random
 import pandas as pd
 from pytest import fixture
-from ds_toolbox.spark_utils import start_local_spark
+from ds_toolbox.utils import start_local_spark
 
 
 @fixture
